Initial commit
Present main work is being done on configloader.py which is a
configuration-type agnostic class that will allow you to use either
environment variables, .ini, or .toml configurations to operate the
main.py code.

The purpose of so many options is to allow the user to run this in
whatever situation/environment they wish. I may add YAML and maybe
JSON configurations, but these will be well-defined with .example
files.